03-完整工程/requirements.txt
transformers>=4.30.0
peft>=0.5.0
datasets>=2.14.0
accelerate>=0.20.0
torch>=2.0.0
pyyaml>=6.0
wandb>=0.15.0  # 可选，用于实验追踪
